Feature/frontend

See merge request s03-final/s03p31b104!20

+++ frontend/src/router/index.js
@@ -78,12 +78,41 @@ const routes = [
   {
     path: '/logout',
     beforeEnter: (to, from, next) => {
-      store.state.email = null;
-      store.state.isLogin = false;
-      localStorage.clear();
-      next('/');
+      Vue.swal({
+        icon: 'warning',
+        title: 'Are You Sure?',
+        showConfirmButton: true,
+        showCancelButton: true,
+      }).then((res) => {
+        if (res.isConfirmed) {
+          setTimeout(() => {
+            store.state.email = null;
+            store.state.isLogin = false;
+            localStorage.clear();
+            next('/');
+          },
+          1500);
+        } else {
+          next(from);
+        }
+      });
+
+      // Swal.fire({
+      //   position: 'top-end',
+      //   icon: 'success',
+      //   title: 'Your work has been saved',
+      //   showConfirmButton: false,
+      //   timer: 1500
+      // })
     },
   },
+  // {
+  //   path: '*',
+  //   beforeEnter: (to, from, next) => {
+  //     alert('Not exist Page');
+  //     next('/');
+  //   },
+  // },
 ];
 
 const router = new VueRouter({
